requirements.txt
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.9.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap4==2.2.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-heroku==0.3.1
django-import-export==2.3.0
django-phonenumber-field==4.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.4
phonenumbers==8.12.6
psycopg2==2.8.5
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
